Worked on simple user interface.

+++ greidinet/__init__.py
@@ -12,18 +12,23 @@ def create_app(configfile=None, default_settings=True):
     from greidinet.model import db
     db.init_app(app)
     # DEBUG
-    app.test_request_context().push()
-    db.drop_all()
-    db.create_all()
+    if (app.debug):
+        app.test_request_context().push()
+        db.drop_all()
+        db.create_all()
     # Use Bootstrap
     Bootstrap(app)
-    # Import blueprints
-    #from greidinet.simplegui import simplegui
-    #app.register_blueprint(simplegui)
+    # Login module
+    from greidinet.login import login_manager, logingui
+    login_manager.init_app(app)
+    app.register_blueprint(logingui)
     # REST api
     from greidinet.restapi import restapi
     restapi.init_app(app, flask_sqlalchemy_db=db)
     restapi.create_apis()
-    
+    # Import blueprints
+    from greidinet.simplegui import simplegui
+    app.register_blueprint(simplegui)
+   
 
     return app
